requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langchain-huggingface==0.0.3
sentence-transformers==2.7.0
qdrant-client==1.10.1
pymupdf==1.24.7
groq==0.9.0
python-dotenv==1.0.1

